+++ src/services/api.ts
@@ -1,12 +1,12 @@
 import 'isomorphic-fetch';
 import fs from 'fs';
 import semver from 'semver';
+import open from 'open';
 import mkdirp from 'mkdirp';
 import yaml from 'yaml';
 import path from 'path';
-import packageJson from 'package-json';
-import { ComponentConfig } from '../types';
-import { getCachedComponentsDirectory } from '../helpers';
+import { ComponentConfig, PackageJson } from '../types';
+import { getCachedComponentsDirectory, getSeconds, sleep } from '../helpers';
 import OptionsService from './options';
 import LoggerService from './logger';
 import ConfigurationService from './configuration';
@@ -27,21 +27,98 @@ export default class ApiService {
     return ApiService.instance;
   }
 
-  async getVfCoreLatestReleaseVersion(): Promise<string> {
-    const options = this.optionsService.getOptions();
+  async authenticateGitHub(): Promise<string> {
+    // (1) get device and user code
+    const response = await fetch('https://github.com/login/device/code', {
+      method: 'POST',
+      body: JSON.stringify({
+        client_id: 'd7d227d46d4ee4e5991a',
+      }),
+      headers: {
+        'Content-Type': 'application/json',
+        Accept: 'application/json',
+      },
+    });
 
-    if (!this.configurationService.config || options.forceRun) {
-      // TODO: handle error, authentication, and add typing
-      const response = await fetch('https://api.github.com/repos/visual-framework/vf-core/tags');
-      const tags: any[] = await response.json();
-
-      this.configurationService.update('vfCoreVersion', tags[0].name);
+    if (!response.ok) {
+      // TODO: use app error
+      throw new Error('Authentication error.');
     }
 
-    return this.configurationService.config.vfCoreVersion;
+    const { user_code, verification_uri, interval, device_code, expires_in } = await response.json();
+    const expiry = Math.floor(expires_in / 60);
+
+    // (2) get access token
+    this.logger.info(`Please enter the code ${user_code} at ${verification_uri}. This expires in ${expiry} minutes.`);
+
+    await sleep(1500);
+    await open(verification_uri);
+
+    let done = false;
+    let lastTimestamp = getSeconds();
+    let accessTokenResponse: Response | null = null;
+    let accessTokenData: any;
+
+    while (!done) {
+      const currentTimestamp = getSeconds();
+
+      if (currentTimestamp - lastTimestamp >= interval + 1) {
+        // this.logger.debug('Making an access token request');
+
+        accessTokenResponse = await fetch('https://github.com/login/oauth/access_token', {
+          method: 'POST',
+          body: JSON.stringify({
+            client_id: 'd7d227d46d4ee4e5991a',
+            device_code: device_code,
+            grant_type: 'urn:ietf:params:oauth:grant-type:device_code',
+          }),
+          headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json',
+          },
+        });
+
+        accessTokenData = await accessTokenResponse.json();
+
+        if (accessTokenResponse.ok && !accessTokenData.error) {
+          this.logger.info('Authentication completed');
+          done = true;
+        } else if (accessTokenData.error === 'access_denied' || accessTokenData.error === 'expired_token') {
+          this.logger.error(accessTokenData.error_description);
+          done = true;
+        } else if (accessTokenData.error === 'authorization_pending') {
+          this.logger.debug(accessTokenData.error_description);
+        }
+
+        lastTimestamp = getSeconds();
+      }
+    }
+
+    return accessTokenData?.access_token;
   }
 
-  async getComponentPackageJson(name: string): Promise<packageJson.AbbreviatedMetadata> {
+  async getVfCoreLatestReleaseVersion(): Promise<string> {
+    // TODO: handle error and add typing
+    const accessToken = this.configurationService.config.gitHubAccessToken;
+    const response = await fetch('https://api.github.com/repos/visual-framework/vf-core/tags', {
+      headers: {
+        Authorization: `token ${accessToken}`,
+      },
+    });
+    const tags: any[] = await response.json();
+    tags.sort(semver.compare);
+
+    return tags[0].name;
+  }
+
+  async getComponentPackageJson(name: string): Promise<PackageJson> {
+    const vfCoreLatestReleaseVersion = this.configurationService.config.vfCoreVersion;
+
+    if (!vfCoreLatestReleaseVersion) {
+      // TODO: define error
+      throw new Error('');
+    }
+
     const options = this.optionsService.getOptions();
     const cachedContentFileName = getCachedComponentsDirectory(name, 'package.json');
 
@@ -54,16 +131,22 @@ export default class ApiService {
 
     this.logger.debug(`Retrieving ${name} package.json from remote`);
 
-    const pkg = await packageJson(`@visual-framework/${name}`);
+    const response = await fetch(this.buildVfCoreContentUrl(vfCoreLatestReleaseVersion, name, 'package.json'));
+    const packageJson: PackageJson = await response.json();
 
     await mkdirp(path.dirname(cachedContentFileName));
-    fs.writeFileSync(cachedContentFileName, JSON.stringify(pkg), 'utf-8');
+    fs.writeFileSync(cachedContentFileName, JSON.stringify(packageJson), 'utf-8');
 
-    return pkg;
+    return packageJson;
   }
 
   async getComponentConfig(name: string): Promise<ComponentConfig> {
-    const vfCoreLatestReleaseVersion = await this.getVfCoreLatestReleaseVersion();
+    const vfCoreLatestReleaseVersion = this.configurationService.config.vfCoreVersion;
+
+    if (!vfCoreLatestReleaseVersion) {
+      // TODO: define error
+      throw new Error('');
+    }
 
     // YAML configuration
     // ------------------
@@ -122,7 +205,13 @@ export default class ApiService {
   }
 
   async getComponentChangelog(name: string): Promise<string> {
-    const vfCoreLatestReleaseVersion = await this.getVfCoreLatestReleaseVersion();
+    const vfCoreLatestReleaseVersion = this.configurationService.config.vfCoreVersion;
+
+    if (!vfCoreLatestReleaseVersion) {
+      // TODO: define error
+      throw new Error('');
+    }
+
     const options = this.optionsService.getOptions();
     const cachedContentFileName = getCachedComponentsDirectory(name, 'CHANGELOG.md');